move graph classes to types module
